requirements.txt
Django==1.11.8
django-registration-redux==2.0
psycopg2==2.7.3.2
pytz==2017.3
